requirements.txt
Django==5.0
pillow==10.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.0